Further work on SubUnit, adding flashes, adding placeholder effects
Had to combine the simplified OnKilled for subs with the complicated
one, so we can still sink them. That's the first change.

Added placeholder effects, the same for all factions, they only differ
if the unit is above or below water.

Also removed the CreateEffects from unit.lua, since everything aside
from land units already had re-defined OnKilled functions in
defaultunits.lua, so they already used custom explosions. I added in
LandUnit aswell, so land units also use them now, without needing the
stuff in Unit.lua. The rest needs to stay there because of overwriting
existing effects, as noted in the file itself. LandUnit was basically
using just MobileUnit, which uses Unit as a base class, and doesn't
modify much as far as effects go, so that was easy. Not sure why units
with a death anim are not spawning effects though. (Test it yourself: a
tank will spawn effects, but, for example, a Mantis will not.)

I also added some flash effects to a few places, but those work and are
not relevant.

There is a problem with this, it seems that any unit with an animation
does not play the custom effect at all (just the white flash). I tried
hooking WalkingLandUnit, but no dice.)